Addressing comments ( capacity properties as strings/validate units in schema)

+++ cluster-operator/src/main/java/io/strimzi/operator/cluster/model/StorageUtils.java
@@ -23,8 +23,8 @@ public class StorageUtils {
      *
      * @return The equivalent number of bytes in the specified units.
      */
-    public static long parseMemory(String memory, String units) {
-        return parseMemory(memory) / memoryFactor(units);
+    public static double parseMemory(String memory, String units) {
+        return parseMemory(memory) / (float) memoryFactor(units);
     }
 
     /**
